对接学校端部分接口，上传测试

+++ src/main/java/com/hklk/oplatform/controller/teacher/EditClassAndStudentByTeacherController.java
@@ -1,14 +1,16 @@
-package com.hklk.oplatform.controller.school;
+package com.hklk.oplatform.controller.teacher;
 
+import com.hklk.oplatform.comm.LoginTeacher;
 import com.hklk.oplatform.controller.BaseController;
+import com.hklk.oplatform.entity.table.Curriculum;
 import com.hklk.oplatform.entity.table.SClass;
 import com.hklk.oplatform.entity.table.SStudent;
-import com.hklk.oplatform.entity.table.User;
+import com.hklk.oplatform.entity.table.STeacher;
+import com.hklk.oplatform.entity.vo.CurriculumForListVo;
+import com.hklk.oplatform.entity.vo.CurriculumVo;
 import com.hklk.oplatform.entity.vo.PageTableForm;
 import com.hklk.oplatform.entity.vo.SClassVo;
-import com.hklk.oplatform.filter.repo.SchoolLoginRepository;
-import com.hklk.oplatform.service.SClassService;
-import com.hklk.oplatform.service.SStudentService;
+import com.hklk.oplatform.service.*;
 import com.hklk.oplatform.util.ExcelUtils;
 import com.hklk.oplatform.util.StatusCode;
 import com.hklk.oplatform.util.ToolUtil;
@@ -27,11 +29,12 @@ import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Logger;
 
-@SchoolLoginRepository
-@RequestMapping("/editStudent")
+@RequestMapping("/tEditStudent")
 @Controller
-public class EditClassAndStudentController extends BaseController {
+public class EditClassAndStudentByTeacherController extends BaseController {
+
     @Autowired
     SClassService sClassService;
     @Autowired
@@ -39,20 +42,21 @@ public class EditClassAndStudentController extends BaseController {
 
     @ResponseBody
     @RequestMapping("/queryClasses")
-    public String queryClasses(String param,int pageNum, HttpServletRequest request,
+    public String queryClasses(String param, int pageNum, HttpServletRequest request,
                                HttpServletResponse response, HttpSession session) {
-        PageTableForm<SClassVo> pageTableForm = sClassService.queryClasses(param,getLoginSchool(request).getSchoolId(), pageNum, pageSize);
+        PageTableForm<SClassVo> pageTableForm = sClassService.queryClasses(param,getLoginTeacher(request).getSchoolId(), pageNum, pageSize);
         return ToolUtil.buildResultStr(StatusCode.SUCCESS, StatusCode.getStatusMsg(StatusCode.SUCCESS), pageTableForm);
     }
+
 
     @ResponseBody
     @RequestMapping("/addOrUpdateClass")
     public String addOrUpdateClass(SClass sClass, HttpServletRequest request,
                                    HttpServletResponse response, HttpSession session) {
-        Integer schoolId = getLoginSchool(request).getSchoolId();
-        sClass.setSchoolId(1);
+        Integer schoolId = getLoginTeacher(request).getSchoolId();
         SClass temp = sClassService.selectByNameForValidate(sClass.getName(), schoolId);
         if (sClass.getId() == null && temp != null) {
+            System.out.println(getLoginTeacher(request).getNickName() + "老师操作!");
             return ToolUtil.buildResultStr(StatusCode.CLASS_EX, StatusCode.getStatusMsg(StatusCode.CLASS_EX));
         } else if (sClass.getId() != null && temp != null && temp.getId() != sClass.getId()) {
             return ToolUtil.buildResultStr(StatusCode.CLASS_EX, StatusCode.getStatusMsg(StatusCode.CLASS_EX));
@@ -66,8 +70,14 @@ public class EditClassAndStudentController extends BaseController {
     @RequestMapping("/delClass")
     public String delClass(Integer id, HttpServletRequest request,
                            HttpServletResponse response, HttpSession session) {
-        sClassService.deleteByPrimaryKey(id);
-        return ToolUtil.buildResultStr(StatusCode.SUCCESS, StatusCode.getStatusMsg(StatusCode.SUCCESS));
+        LoginTeacher loginTeacher = getLoginTeacher(request);
+        SClass temp = sClassService.selectByPrimaryKey(id);
+        if (temp.getCreateBy() == null || temp.getCreateBy() != loginTeacher.getTeacherId()) {
+            return ToolUtil.buildResultStr(StatusCode.VALIDATE_CLASS_IS_TEACHER_CREATE, StatusCode.getStatusMsg(StatusCode.VALIDATE_CLASS_IS_TEACHER_CREATE));
+        } else {
+            sClassService.deleteByPrimaryKey(id);
+            return ToolUtil.buildResultStr(StatusCode.SUCCESS, StatusCode.getStatusMsg(StatusCode.SUCCESS));
+        }
     }
 
     @ResponseBody
@@ -82,8 +92,9 @@ public class EditClassAndStudentController extends BaseController {
     @RequestMapping("/addOrUpdateStudent")
     public String addOrUpdateStudent(SStudent sStudent, HttpServletRequest request,
                                      HttpServletResponse response, HttpSession session) {
-        SStudent temp = sStudentService.selectBySNumForValidate(getLoginSchool(request).getSchoolId(), sStudent.getsNum());
+        SStudent temp = sStudentService.selectBySNumForValidate(getLoginTeacher(request).getSchoolId(), sStudent.getsNum());
         if (sStudent.getId() == null && temp != null) {
+            System.out.println(getLoginTeacher(request).getNickName() + "老师操作!");
             return ToolUtil.buildResultStr(StatusCode.STUDENT_EX, StatusCode.getStatusMsg(StatusCode.STUDENT_EX));
         } else if (sStudent.getId() != null && temp != null && temp.getId() != sStudent.getId()) {
             return ToolUtil.buildResultStr(StatusCode.STUDENT_EX, StatusCode.getStatusMsg(StatusCode.STUDENT_EX));
@@ -102,6 +113,7 @@ public class EditClassAndStudentController extends BaseController {
         String[] fieldNames = {"学号", "姓名", "性别", "家长姓名", "家长联系电话"};
         List<SStudent> result = sStudentService.queryStudentByClassId(classId);
         try {
+            System.out.println("teacher export ----------------------");
             ServletOutputStream out = response.getOutputStream();
             response.setHeader("Content-disposition", "attachment; filename=students.xlsx");
             response.setContentType("application/msexcel");
@@ -125,20 +137,21 @@ public class EditClassAndStudentController extends BaseController {
             file.transferTo(fileTemp);
             String[] fieldNames = {"sNum", "fullName", "sex", "parentName", "parentPhone"};
             List<SStudent> sStudents = ExcelUtils.importExcel(fileTemp, SStudent.class, fieldNames, 1, 0, 0);
-            Integer schoolId = getLoginSchool(request).getSchoolId();
+            Integer schoolId = getLoginTeacher(request).getSchoolId();
             List<SStudent> errorInsert = new ArrayList<>();
-            int index = 0;
+
+            int inx = 0;
             for (SStudent sStudent : sStudents) {
                 SStudent temp = sStudentService.selectBySNumForValidate(schoolId, sStudent.getsNum());
                 if (sStudent.getId() == null && temp != null) {
                     errorInsert.add(sStudent);
-                    index++;
+                    inx++;
                 } else {
                     sStudent.setClassId(classId);
                     sStudentService.insertOrUpdateByPrimaryKeySelective(sStudent);
                 }
             }
-            if (index > 0) {
+            if (inx > 0) {
                 return ToolUtil.buildResultStr(StatusCode.IMPORTERROR_STUDENT, StatusCode.getStatusMsg(StatusCode.IMPORTERROR_STUDENT), errorInsert);
             } else {
                 return ToolUtil.buildResultStr(StatusCode.SUCCESS, StatusCode.getStatusMsg(StatusCode.SUCCESS));
@@ -148,4 +161,5 @@ public class EditClassAndStudentController extends BaseController {
             return ToolUtil.buildResultStr(StatusCode.ERROR, StatusCode.getStatusMsg(StatusCode.ERROR));
         }
     }
+
 }
